Adding scripts that were made or modified
I created the Coin script for detecting whether or not it has collided with the player object and then modified the PlayerController by adding a GainAPoint public void that would be called by the Coin script. Then I modified the GameManager script by allocating the Coin prefab and script along with the scoring mechanic in order to finish the project. This was done by adding a few void classes regarding creating the coin instance and changing the text score after updating it.